requirements.txt
pyqt
PyQt-tools

